Rename flutter_config package to flutter
